requirements.txt
attrs==21.2.0
blis==0.7.4
catalogue==2.0.4
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
cymem==2.0.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
filelock==3.0.12
Flask==2.0.1
ftfy==6.0.3
huggingface-hub==0.0.12
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
murmurhash==1.0.5
numpy==1.21.0
overrides==6.1.0
packaging==21.0
pathy==0.6.0
Pillow==8.3.0
pluggy==0.13.1
preshed==3.0.5
py==1.10.0
pydantic==1.7.4
pyparsing==2.4.7
pytest==6.2.4
PyYAML==5.4.1
regex==2021.7.6
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
sacremoses==0.0.45
scipy==1.7.0
six==1.16.0
smart-open==5.1.0
spacy==3.0.6
spacy-legacy==3.0.7
srsly==2.4.1
thinc==8.0.7
timm==0.4.12
tokenizers==0.10.3
toml==0.10.2
torchray==1.0.0.2
tqdm==4.61.2
transformers==4.9.0
typer==0.3.2
typing-extensions==3.10.0.0
typing-utils==0.1.0
urllib3==1.26.6
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==2.0.1
